refactor: domain 아키텍쳐 -> layered 아키텍쳐 변경
