+++ src/main/java/our/yurivongella/instagramclone/controller/FeedController.java
@@ -7,7 +7,7 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
-import our.yurivongella.instagramclone.controller.dto.PostResDto;
+import our.yurivongella.instagramclone.controller.dto.post.PostResDto;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
