+++ src/main/java/our/yurivongella/instagramclone/controller/dto/comment/CommentDto.java
@@ -4,13 +4,13 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 
 import lombok.ToString;
 import our.yurivongella.instagramclone.controller.dto.member.MemberDto;
-import our.yurivongella.instagramclone.domain.member.Member;
+import our.yurivongella.instagramclone.entity.Member;
 
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
-import our.yurivongella.instagramclone.domain.comment.Comment;
+import our.yurivongella.instagramclone.entity.Comment;
 
 import java.time.LocalDateTime;
