+++ src/main/java/our/yurivongella/instagramclone/controller/dto/profile/ProfilePostDto.java
@@ -2,7 +2,7 @@ package our.yurivongella.instagramclone.controller.dto.profile;
 
 import lombok.*;
 import our.yurivongella.instagramclone.controller.dto.post.MediaUrlDto;
-import our.yurivongella.instagramclone.domain.post.Post;
+import our.yurivongella.instagramclone.entity.Post;
 import our.yurivongella.instagramclone.util.DateTimeUtil;
 
 import java.util.List;
