+++ src/main/java/our/yurivongella/instagramclone/repository/RefreshTokenRepository.java
@@ -1,7 +1,8 @@
-package our.yurivongella.instagramclone.domain.refeshtoken;
+package our.yurivongella.instagramclone.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
+import our.yurivongella.instagramclone.entity.RefreshToken;
 
 import java.util.Optional;
