+++ src/main/java/our/yurivongella/instagramclone/repository/comment/CommentRepository.java
@@ -1,10 +1,8 @@
-package our.yurivongella.instagramclone.domain.comment;
-
-import java.util.List;
+package our.yurivongella.instagramclone.repository.comment;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import our.yurivongella.instagramclone.domain.post.Post;
+import our.yurivongella.instagramclone.entity.Comment;
 
 public interface CommentRepository extends JpaRepository<Comment, Long> , CustomCommentRepository {
 }
