+++ src/main/java/our/yurivongella/instagramclone/repository/post/PostRepository.java
@@ -1,13 +1,12 @@
-package our.yurivongella.instagramclone.domain.post;
+package our.yurivongella.instagramclone.repository.post;
 
 import java.util.List;
 
 import org.springframework.data.domain.Pageable;
-import org.springframework.data.domain.Slice;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import our.yurivongella.instagramclone.domain.comment.Comment;
-import our.yurivongella.instagramclone.domain.member.Member;
+import our.yurivongella.instagramclone.entity.Member;
+import our.yurivongella.instagramclone.entity.Post;
 
 public interface PostRepository extends JpaRepository<Post, Long>, CustomPostRepository {
     List<Post> findAllByMember(Member member);
