+++ src/test/java/our/yurivongella/instagramclone/controller/dto/MemberDtoTest.java
@@ -3,8 +3,8 @@ package our.yurivongella.instagramclone.controller.dto;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import our.yurivongella.instagramclone.controller.dto.member.MemberDto;
-import our.yurivongella.instagramclone.domain.follow.Follow;
-import our.yurivongella.instagramclone.domain.member.Member;
+import our.yurivongella.instagramclone.entity.Follow;
+import our.yurivongella.instagramclone.entity.Member;
 
 import static org.assertj.core.api.Assertions.*;
